# Conflicts:
#	Source/com/drew/metadata/xmp/XmpDirectory.java
#	Source/com/drew/metadata/xmp/XmpReader.java
#	Source/com/drew/tools/ProcessAllImagesInFolderUtility.java
#	Tests/com/drew/metadata/xmp/XmpReaderTest.java
#	pom.xml

+++ Source/com/drew/tools/ProcessAllImagesInFolderUtility.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2016 Drew Noakes
+ * Copyright 2002-2019 Drew Noakes and contributors
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -24,7 +24,10 @@ package com.drew.tools;
 import com.adobe.internal.xmp.XMPException;
 import com.adobe.internal.xmp.XMPIterator;
 import com.adobe.internal.xmp.XMPMeta;
+import com.adobe.internal.xmp.options.IteratorOptions;
 import com.adobe.internal.xmp.properties.XMPPropertyInfo;
+import com.drew.imaging.FileType;
+import com.drew.imaging.FileTypeDetector;
 import com.drew.imaging.ImageMetadataReader;
 import com.drew.imaging.jpeg.JpegProcessingException;
 import com.drew.lang.StringUtil;
@@ -36,7 +39,7 @@ import com.drew.metadata.Tag;
 import com.drew.metadata.exif.ExifIFD0Directory;
 import com.drew.metadata.exif.ExifSubIFDDirectory;
 import com.drew.metadata.exif.ExifThumbnailDirectory;
-import com.drew.metadata.file.FileMetadataDirectory;
+import com.drew.metadata.file.FileSystemDirectory;
 import com.drew.metadata.xmp.XmpDirectory;
 
 import java.io.*;
@@ -47,7 +50,7 @@ import java.util.*;
  */
 public class ProcessAllImagesInFolderUtility
 {
-    public static void main(String[] args) throws IOException, JpegProcessingException
+    public static void main(String[] args) throws IOException
     {
         List<String> directories = new ArrayList<String>();
 
@@ -168,11 +171,16 @@ public class ProcessAllImagesInFolderUtility
 
     abstract static class FileHandlerBase implements FileHandler
     {
+        // TODO obtain these from FileType enum directly
         private final Set<String> _supportedExtensions = new HashSet<String>(
             Arrays.asList(
-                "jpg", "jpeg", "png", "gif", "bmp", "ico", "webp", "pcx", "ai", "eps",
-                "nef", "crw", "cr2", "orf", "arw", "raf", "srw", "x3f", "rw2", "rwl",
-                "tif", "tiff", "psd", "dng"));
+                "jpg", "jpeg", "png", "gif", "bmp", "heic", "heif", "ico", "webp", "pcx", "ai", "eps",
+                "nef", "crw", "cr2", "orf", "arw", "raf", "srw", "x3f", "rw2", "rwl", "dcr",
+                "tif", "tiff", "psd", "dng",
+                "mp3",
+                "j2c", "jp2", "jpf", "jpm", "mj2",
+                "3g2", "3gp", "m4v", "mov", "mp4", "m2v", "mts",
+                "pbm", "pnm", "pgm", "ppm"));
 
         private int _processedFileCount = 0;
         private int _exceptionCount = 0;
@@ -253,23 +261,26 @@ public class ProcessAllImagesInFolderUtility
             super.onStartingDirectory(directoryPath);
 
             // Delete any existing 'metadata' folder
-            File metadataDirectory = new File(directoryPath + "/metadata");
+            File metadataDirectory = new File(directoryPath + "/metadata/java");
             if (metadataDirectory.exists())
                 deleteRecursively(metadataDirectory);
         }
 
-        private static void deleteRecursively(File directory)
+        private static void deleteRecursively(@NotNull File directory)
         {
             if (!directory.isDirectory())
                 throw new IllegalArgumentException("Must be a directory.");
 
             if (directory.exists()) {
-                for (String item : directory.list()) {
-                    File file = new File(item);
-                    if (file.isDirectory())
-                        deleteRecursively(file);
-                    else
-                        file.delete();
+                String[] list = directory.list();
+                if (list != null) {
+                    for (String item : list) {
+                        File file = new File(item);
+                        if (file.isDirectory())
+                            deleteRecursively(file);
+                        else
+                            file.delete();
+                    }
                 }
             }
 
@@ -312,11 +323,16 @@ public class ProcessAllImagesInFolderUtility
                         // Write the directory's tags
                         for (Tag tag : directory.getTags()) {
                             String tagName = tag.getTagName();
-                            String description = tag.getDescription();
+                            String description;
+                            try {
+                                description = tag.getDescription();
+                            } catch (Exception ex) {
+                                description = "ERROR: " + ex.getMessage();
+                            }
                             if (description == null)
                                 description = "";
                             // Skip the file write-time as this changes based on the time at which the regression test image repository was cloned
-                            if (directory instanceof FileMetadataDirectory && tag.getTagType() == FileMetadataDirectory.TAG_FILE_MODIFIED_DATE)
+                            if (directory instanceof FileSystemDirectory && tag.getTagType() == FileSystemDirectory.TAG_FILE_MODIFIED_DATE)
                                 description = "<omitted for regression testing as checkout dependent>";
                             writer.format("[%s - %s] %s = %s%s", directoryName, tag.getTagTypeHex(), tagName, description, NEW_LINE);
                         }
@@ -324,20 +340,34 @@ public class ProcessAllImagesInFolderUtility
                             writer.write(NEW_LINE);
                         // Special handling for XMP directory data
                         if (directory instanceof XmpDirectory) {
-                            Collection<XmpDirectory> xmpDirectories = metadata.getDirectoriesOfType(XmpDirectory.class);
                             boolean wrote = false;
-                            for (XmpDirectory xmpDirectory : xmpDirectories) {
-                                XMPMeta xmpMeta = xmpDirectory.getXMPMeta();
-                                try {
-                                    XMPIterator iterator = xmpMeta.iterator();
-                                    while (iterator.hasNext()) {
-                                        XMPPropertyInfo prop = (XMPPropertyInfo)iterator.next();
-                                        writer.format("[XMPMeta - %s] %s = %s%s", prop.getNamespace(), prop.getPath(), prop.getValue(), NEW_LINE);
-                                        wrote = true;
-                                    }
-                                } catch (XMPException e) {
-                                    e.printStackTrace();
+                            XmpDirectory xmpDirectory = (XmpDirectory)directory;
+                            XMPMeta xmpMeta = xmpDirectory.getXMPMeta();
+                            try {
+                                IteratorOptions options = new IteratorOptions().setJustLeafnodes(true);
+                                XMPIterator iterator = xmpMeta.iterator(options);
+                                while (iterator.hasNext()) {
+                                    XMPPropertyInfo prop = (XMPPropertyInfo)iterator.next();
+                                    String ns = prop.getNamespace();
+                                    String path = prop.getPath();
+                                    String value = prop.getValue();
+
+                                    if (path == null)
+                                        continue;
+                                    if (ns == null)
+                                        ns = "";
+
+                                    final int MAX_XMP_VALUE_LENGTH = 512;
+                                    if (value == null)
+                                        value = "";
+                                    else if (value.length() > MAX_XMP_VALUE_LENGTH)
+                                        value = String.format("%s <truncated from %d characters>", value.substring(0, MAX_XMP_VALUE_LENGTH), value.length());
+
+                                    writer.format("[XMPMeta - %s] %s = %s%s", ns, path, value, NEW_LINE);
+                                    wrote = true;
                                 }
+                            } catch (XMPException e) {
+                                e.printStackTrace();
                             }
                             if (wrote)
                                 writer.write(NEW_LINE);
@@ -405,13 +435,29 @@ public class ProcessAllImagesInFolderUtility
             if (!metadataDir.exists())
                 metadataDir.mkdir();
 
-            String outputPath = String.format("%s/metadata/%s.txt", file.getParent(), file.getName());
+            File javaDir = new File(String.format("%s/metadata/java", file.getParent()));
+            if (!javaDir.exists())
+                javaDir.mkdir();
+
+            String outputPath = String.format("%s/metadata/java/%s.txt", file.getParent(), file.getName());
             Writer writer = new OutputStreamWriter(
                 new FileOutputStream(outputPath),
                 "UTF-8"
             );
             writer.write("FILE: " + file.getName() + NEW_LINE);
-            writer.write(NEW_LINE);
+
+            // Detect file type
+            BufferedInputStream stream = null;
+            try {
+                stream = new BufferedInputStream(new FileInputStream(file));
+                FileType fileType = FileTypeDetector.detectFileType(stream);
+                writer.write(String.format("TYPE: %s" + NEW_LINE, fileType.toString().toUpperCase()));
+                writer.write(NEW_LINE);
+            } finally {
+                if (stream != null) {
+                    stream.close();
+                }
+            }
 
             return new PrintWriter(writer);
         }
@@ -435,7 +481,7 @@ public class ProcessAllImagesInFolderUtility
         private final Map<String, String> _extensionEquivalence = new HashMap<String, String>();
         private final Map<String, List<Row>> _rowListByExtension = new HashMap<String, List<Row>>();
 
-        class Row
+        static class Row
         {
             final File file;
             final Metadata metadata;
@@ -543,12 +589,14 @@ public class ProcessAllImagesInFolderUtility
             Writer writer = new OutputStreamWriter(stream);
             writer.write("# Image Database Summary\n\n");
 
-            for (String extension : _rowListByExtension.keySet()) {
+            for (Map.Entry<String, List<Row>> entry : _rowListByExtension.entrySet()) {
+                String extension = entry.getKey();
                 writer.write("## " + extension.toUpperCase() + " Files\n\n");
 
                 writer.write("File|Manufacturer|Model|Dir Count|Exif?|Makernote|Thumbnail|All Data\n");
                 writer.write("----|------------|-----|---------|-----|---------|---------|--------\n");
-                List<Row> rows = _rowListByExtension.get(extension);
+
+                List<Row> rows = entry.getValue();
 
                 // Order by manufacturer, then model
                 Collections.sort(rows, new Comparator<Row>() {
